Remove need to have dummy injections of wear dependencies in phone app
